refactor: Se reorganizan los paquetes para evitar problemas de importación cíclica
