+++ shared/infrastructure/validation/validation.go
@@ -1,3 +1,5 @@
+// TODO: Pendientes tests
+
 package validation
 
 import (
