+++ user/infrastructure/repository/gorm.go
@@ -1,3 +1,5 @@
+// TODO: Pendientes tests
+
 package repository
 
 import (
